feat: 필터에 따른 모임 노출 변경 및 customhooks 추가, modify: 헤더 텍스트 수정, 모임 리스트 로직 변수 수정
